+++ src/asgard_term_sources.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
 #include "asgard_interp.hpp"
-#include "asgard_moment.hpp"
+#include "asgard_moment_manager.hpp"
 
 namespace asgard
 {
